+++ examples/resources/datarobot_deployment/resource.tf
@@ -45,6 +45,7 @@ resource "datarobot_deployment" "example" {
       value = "val",
     },
   ]
+  retraining_settings                   = {}
 }
 
 output "datarobot_deployment_id" {